Add CRUDService<T> interface
As we all write our services separately, we should at least make sure that their APIs match.
The easiest way to do it is to make every CRUD service implement the CRUDService interface with TEntity = the DTO class they perform operations on.
Every method in this interface throws two types of errors:
1. IllegalArgumentException - should be thrown when the given data is incorrect (out of range, doesn't follow limitations and so on)
2. IOException - should be thrown when there's an issue not with the arguments user provided, but with the database itself (access issue or anything else that is thrown by the database)